+++ blob.go
@@ -5,7 +5,7 @@ package dom
 import "syscall/js"
 
 type BlobIFace interface {
-	GetSize() float64
+	GetSize() int
 	Slice(args ...interface{}) Blob
 	GetType() string
 }
@@ -15,9 +15,9 @@ type Blob struct {
 
 func JSValueToBlob(val js.Value) Blob { return Blob{Value: Value{Value: val}} }
 func (v Value) AsBlob() Blob          { return Blob{Value: v} }
-func (b Blob) GetSize() float64 {
+func (b Blob) GetSize() int {
 	val := b.Get("size")
-	return val.Float()
+	return val.Int()
 }
 func (b Blob) Slice(args ...interface{}) Blob {
 	val := b.Call("slice", args...)
